test(anyharness): split config restore coverage
