+++ anyharness/crates/anyharness-lib/src/live/sessions/actor/tests/mod.rs
@@ -49,6 +49,7 @@ use tokio::sync::{broadcast, mpsc, Mutex, RwLock};
 
 mod conditional_cancel;
 mod config;
+mod config_restore;
 mod config_variants;
 mod domain_ops;
 mod notifications;
